medisure-ai-backend-4874b60d9380591b433e899995d5cce6c3773183/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2; python_version >= "3.5"
chardet==3.0.4
click==7.1.2
flask==1.1.2
flask-cors==3.0.9
google-api-core==1.22.2
google-auth==1.21.1
google-cloud-core==1.4.1
google-cloud-documentai==0.2.0
google-cloud-storage==1.31.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0; python_version >= "3.5"
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
numpy==1.19.2
openai==0.2.4
proto-plus==1.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20; python_version >= "3.5"
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
rsa==4.6; python_version >= "3.5"
scipy==1.5.2
six==1.15.0
summa==1.2.0
urllib3==1.25.10
werkzeug==1.0.1
